Revert "name_extract_API.py 위치 이동 후 최신화"
This reverts commit 07e9d81961764005d7b6f05afb897d15f31d0a57.
